\Total:                   70.0 of 70.0\

+++ read-line.cc
@@ -94,7 +94,7 @@ char *read_line() {
       line_pos = 0;
     } else if (in_char == 4) { // <ctrl-D> / delete key - delete character
       // Only delete if the line length is longer or equal to 0
-      if ((line_pos >= 0)) {   // Don't delete last
+      if ((line_pos >= 0) && (line_pos != total_chars)) { // Don't delete last
         delete_char(line_pos); // Delete the character
         total_chars--;
